requirements_blog.txt
# Mini Golf Every Day - Blog Backend Requirements
# Core Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication and security
PyJWT==2.8.0
bcrypt==4.1.2
Werkzeug==3.0.1

# Data processing
requests==2.31.0

# Development
python-dotenv==1.0.0

# Production WSGI server
gunicorn==21.2.0

# Existing requirements for video functionality
yt-dlp==2023.12.30
